requirements.txt
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-instrumentation-flask==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation==0.40b0
opentelemetry-semantic-conventions==0.40b0
opentelemetry-instrumentation-aws-lambda==0.40b0
opentelemetry-instrumentation-dbapi==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-instrumentation-sqlite3==0.40b0
opentelemetry-instrumentation-urllib==0.40b0
opentelemetry-instrumentation-wsgi==0.40b0
opentelemetry-instrumentation-flask==0.40b0
opentelemetry-instrumentation-grpc==0.40b0
opentelemetry-instrumentation-jinja2==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-system-metrics==0.40b0
opentelemetry-instrumentation-tornado==0.40b0
opentelemetry-instrumentation-urllib3==0.40b0


Flask
Requests==2.31.0
six==1.16.0
python-dotenv
protobuf==4.23.*
beeline